+++ modules/tracer/frontend/src/main/scala/Frontend.scala
@@ -19,43 +19,6 @@ import scala.scalajs.js.Date
 import scala.scalajs.js.JSON
 import scala.scalajs.js.annotation.JSGlobal
 
-object Api:
-
-  given JsonValueCodec[Vector[Message]]        = JsonCodecMaker.make
-  given rw: JsonValueCodec[Vector[RawMessage]] = JsonCodecMaker.make
-
-  def all: Future[Vector[Message]] =
-    fetch("/api/all")
-      .flatMap(_.text())
-      .map(str => readFromStringReentrant[Vector[Message]](str))
-
-  def rawAll: Future[Vector[RawMessage]] =
-    fetch("/api/raw/all")
-      .flatMap(_.text())
-      .map(str => readFromStringReentrant[Vector[RawMessage]](str))
-
-  def request(id: String): Future[Option[RawMessage]] =
-    fetch(s"/api/raw/request/$id")
-      .flatMap(resp =>
-        if resp.status == 200 then
-          resp
-            .text()
-            .map(readFromStringReentrant[RawMessage](_))
-            .map(Option.apply)
-        else Future.successful(None)
-      )
-  def response(id: String): Future[Option[RawMessage]] =
-    fetch(s"/api/raw/response/$id")
-      .flatMap(resp =>
-        if resp.status == 200 then
-          resp
-            .text()
-            .map(readFromStringReentrant[RawMessage](_))
-            .map(Option.apply)
-        else Future.successful(None)
-      )
-end Api
-
 @js.native
 @JSGlobal
 object hljs extends js.Object:
@@ -213,9 +176,9 @@ object Frontend:
             (cm.params, cm.error) match
               case (None, None) => i("no error or payload")
               case (_, Some(err)) =>
-                div(b(color := "pink", "Error"), displayJson(err))
-              case (Some(p), None) =>
-                div(b(color := "lightgreen", "Params"), displayJson(p))
+                displayErr(err)
+              case (p, None) =>
+                displayPayload("Params", p)
         }
       ),
       div(
